+++ mediaplatform/admin.py
@@ -7,7 +7,8 @@ from django.urls import reverse
 from django.utils.formats import localize
 from django.utils.html import format_html
 
-from mediaplatform_jwp import models as jwpmodels, jwplatform as api
+from mediaplatform_jwp import models as jwpmodels
+from mediaplatform_jwp.jwplatformapi import deliveryapi as api
 from legacysms import models as smsmodels
 
 from . import models
